# Add ARM version of calculating mode scores