### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description

This PR adds a ARM version of calculating mode scores which is used during webp encoding. Implementation is based on [libwebp/enc_neon.c](https://github.com/webmproject/libwebp/blob/main/src/dsp/enc_neon.c)

Benchmarks:

main
```
BenchmarkDotNet=v0.13.0, OS=ubuntu 20.04
Unknown processor
.NET SDK=6.0.405
  [Host]     : .NET 6.0.13 (6.0.1322.58009), Arm64 RyuJIT
  Job-QPBFHS : .NET 6.0.13 (6.0.1322.58009), Arm64 RyuJIT

Runtime=.NET 6.0  Arguments=/p:DebugType=portable  IterationCount=3
LaunchCount=1  WarmupCount=3

|                     Method |                   TestImage |       Mean |     Error |   StdDev | Ratio |      Gen 0 |     Gen 1 |     Gen 2 |  Allocated |
|--------------------------- |---------------------------- |-----------:|----------:|---------:|------:|-----------:|----------:|----------:|-----------:|
|        'Magick Webp Lossy' | Jpg/baseline/Calliphora.jpg |   367.1 ms |   2.72 ms |  0.15 ms |  0.19 |          - |         - |         - |     530 KB |
|    'ImageSharp Webp Lossy' | Jpg/baseline/Calliphora.jpg | 2,962.1 ms |  77.63 ms |  4.26 ms |  1.56 | 27000.0000 | 3000.0000 | 1000.0000 |  71,752 KB |
```

PR
```
|                     Method |                   TestImage |       Mean |     Error |   StdDev | Ratio |      Gen 0 |     Gen 1 |     Gen 2 |  Allocated |
|--------------------------- |---------------------------- |-----------:|----------:|---------:|------:|-----------:|----------:|----------:|-----------:|
|        'Magick Webp Lossy' | Jpg/baseline/Calliphora.jpg |   377.0 ms | 311.58 ms | 17.08 ms |  0.20 |          - |         - |         - |     530 KB |
|    'ImageSharp Webp Lossy' | Jpg/baseline/Calliphora.jpg | 2,830.9 ms |  22.73 ms |  1.25 ms |  1.49 | 27000.0000 | 3000.0000 | 1000.0000 |  71,752 KB |
```

Test image was ` Jpg/baseline/Calliphora.jpg` from the tests/Images/Input folder.

cpu info
```
Architecture:        aarch64
CPU op-mode(s):      32-bit, 64-bit
Byte Order:          Little Endian
CPU(s):              6
On-line CPU(s) list: 0-5
Thread(s) per core:  1
Core(s) per socket:  3
Socket(s):           2
Vendor ID:           ARM
Model:               4
Model name:          Cortex-A53
Stepping:            r0p4
CPU max MHz:         1896.0000
CPU min MHz:         100.0000
BogoMIPS:            48.00
Flags:               fp asimd evtstrm aes pmull sha1 sha2 crc32
```